+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+ - Groups: We are displaying the groups you are a member of on your profile.
  - Profile: Made profiles just a bit nicer!
  - Navigation: We fixed the context menu for the back button.
  - Notifications: Fixed a little layout issue in the notifications screen.